Define a non-generic base class for LanguageOperationResult that contains only success & messages.
